feat(extensions): split footer rendering into agent-footer baseline
Move the custom footer (sandbox dir + token stats + extension statuses)
out of the sandbox safety extension into a dedicated `agent-footer`
extension and add a tools list opposite the sandbox dir. Line 1 of the
footer now shows `sandbox: <root>` left-aligned and
`tools: <name1, name2, ...>` (from pi.getActiveTools()) right-aligned,
truncating the tools list with an ellipsis on narrow terminals.

sandbox.ts is back to safety-only (bash block + path clamp) and
agent-footer joins sandbox / no-startup-help as the third baseline
extension in BASELINE_EXTENSIONS.
